Extract wanted persistence into repository mixin
