+++ changelog.d/+wanted-repository-slice.changed.md
@@ -1,0 +1,2 @@
+Moved wanted hook and wanted interest persistence into a dedicated repository
+slice while preserving the public forum repository facade.
